+++ src/client.ts
@@ -2,8 +2,8 @@
 // Exports
 //
 
+export * from "./components/forms/Control.client.js";
 export * from "./components/forms/Form.client.js";
-export * from "./components/forms/Input.client.js";
 
 export * from "./components/HumanDateTime.client.js";
 export * from "./components/HumanRelativeTime.client.js";