+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 * Removed the `Select`, `SelectOption` and `SelectOptionGroup` components.
 * Removed the `TextArea` component.
 * Removed the `TextInput` component.
-* Added the `Input` component. This is a generic input component that can be used to create several types of inputs.
+* Added the `Control` component. This is a generic input component that can be used to create several types of inputs.
 * Added an optional options object to the `Block` component.
 * Removed `BlockWithoutPadding` component.
 * Changed options for the `Button` component.